+++ CHANGES.md
@@ -1,3 +1,8 @@
-# 2.1.2
+### 2.1.3
+ * Upgrades deps
+ * Adds `->event`, `dispatch-event`, and `o-merge!` to `c3kit.wire.js`
+ * Passes cljs specs under `:advanced` optimizations
+
+### 2.1.2
  * using apron 2.1.3, replace use of deprecated schema fns
  * update from ring/ring 1.10.0 to 1.12.0